tune parameters

+++ analysis/check.py
@@ -33,7 +33,7 @@ epoch = 10
 with torch.autograd.detect_anomaly():
 	for e in range(epoch):
 		i = random.randint(0, 3)
-		data, label = torch.tensor(dataset[i]), torch.tensor(labelset[i], dtype=torch.long)
+		data, label = torch.tensor(dataset[i], dtype=torch.long), torch.tensor(labelset[i], dtype=torch.long)
 
 		data = torch.autograd.Variable(data).cuda()
 		label = torch.autograd.Variable(label).cuda()
